+++ src/main/java/com/ashwani/HealthCare/Repository/PaymentRepository.java
@@ -1,9 +1,9 @@
 package com.ashwani.HealthCare.Repository;
 
-import com.ashwani.HealthCare.Entity.Payment;
+import com.ashwani.HealthCare.Entity.PaymentEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-    Payment findByOrderId(String orderId);
+public interface PaymentRepository extends JpaRepository<PaymentEntity, Long> {
+    PaymentEntity findByOrderId(String orderId);
 }